# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
  description: 'TaskY API — canonical contract generated from springdoc on a dev-profile boot (Observed). Regenerate: boot
    api with dev profile, GET /api-docs, convert to YAML.'
  x-generated: 2026-09-26, feat OIDC code exchange + timer e2e
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/v1/requests/{requestId}:
    get:
      tags:
      - internal-request-controller
      operationId: get
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/InternalRequestResponse'
    put:
      tags:
      - internal-request-controller
      operationId: update_1
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInternalRequestRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/InternalRequestResponse'
    delete:
      tags:
      - internal-request-controller
      operationId: delete_1
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/requests:
    get:
      tags:
      - internal-request-controller
      operationId: list_2
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: priority
        in: query
        required: false
        schema:
          type: string
      - name: assigneeId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: responsibleDepartmentId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: mine
        in: query
        required: false
        schema:
          type: boolean
          default: false
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 100
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginatedResponseInternalRequestResponse'
    post:
      tags:
      - internal-request-controller
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInternalRequestRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/InternalRequestResponse'
  /api/v1/requests/{requestId}/tasks:
    get:
      tags:
      - internal-request-controller
      operationId: tasks
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivityResponse'
    post:
      tags:
      - internal-request-controller
      operationId: createTasks
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequestTasksRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivityResponse'
  /api/v1/requests/{requestId}/convert-to-project:
    post:
      tags:
      - internal-request-controller
      operationId: convertToProject
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConvertRequestToProjectRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/InternalRequestResponse'
  /api/v1/requests/{requestId}/comments:
    get:
      tags:
      - internal-request-controller
      operationId: comments
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RequestCommentResponse'
    post:
      tags:
      - internal-request-controller
      operationId: addComment
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequestCommentRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RequestCommentResponse'
...
  /api/v1/activities/{activityId}/comments:
    get:
      tags:
      - activity-controller
      operationId: comments_1
      parameters:
      - name: activityId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivityCommentResponse'
    post:
      tags:
      - activity-controller
      operationId: addComment_1
      parameters:
      - name: activityId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActivityCommentRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ActivityCommentResponse'
...
  /api/v1/requests/{requestId}/status:
    patch:
      tags:
      - internal-request-controller
      operationId: changeStatus
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeRequestStatusRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/InternalRequestResponse'
  /api/v1/requests/{requestId}/link-project:
    patch:
      tags:
      - internal-request-controller
      operationId: linkProject
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkRequestProjectRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/InternalRequestResponse'
  /api/v1/requests/{requestId}/link-activity:
    patch:
      tags:
      - internal-request-controller
      operationId: linkActivity
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkRequestActivityRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/InternalRequestResponse'
  /api/v1/requests/{requestId}/assign:
    patch:
      tags:
      - internal-request-controller
      operationId: assign_1
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRequestRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/InternalRequestResponse'
...
  /api/v1/requests/{requestId}/comments/{commentId}:
    delete:
      tags:
      - internal-request-controller
      operationId: deleteComment
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: commentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/activities/{activityId}/comments/{commentId}:
    delete:
      tags:
      - activity-controller
      operationId: deleteComment_1
      parameters:
      - name: activityId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: commentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
...
components:
  schemas:
...
    UpdateInternalRequestRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 0
        description:
          type: string
          maxLength: 5000
          minLength: 0
        glpiTicketId:
          type: string
          maxLength: 64
          minLength: 0
        priority:
          type: string
        responsibleDepartmentId:
          type: string
          format: uuid
        assigneeMembershipId:
          type: string
          format: uuid
        assigneeMembershipIds:
          type: array
          items:
            type: string
            format: uuid
        desiredDueDate:
          type: string
          format: date-time
    InternalRequestResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        requestKey:
          type: string
        glpiTicketId:
          type: string
        title:
          type: string
        description:
          type: string
        priority:
          type: string
        status:
          type: string
        requesterMembershipId:
          type: string
          format: uuid
        requestingDepartmentId:
          type: string
          format: uuid
        responsibleDepartmentId:
          type: string
          format: uuid
        assigneeMembershipId:
          type: string
          format: uuid
        assigneeMembershipIds:
          type: array
          items:
            type: string
            format: uuid
        desiredDueDate:
          type: string
          format: date-time
        projectId:
          type: string
          format: uuid
        activityId:
          type: string
          format: uuid
        completedAt:
          type: string
          format: date-time
        canceledAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ActivityResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        parentActivityId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        weight:
          type: integer
          format: int32
        startDatetime:
          type: string
          format: date-time
        endDatetime:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - TODO
          - IN_PROGRESS
          - IN_TESTING
          - DONE
          - BLOCKED
          - CANCELED
        taskType:
          type: string
          enum:
          - TASK
          - BUG
          - IMPROVEMENT
          - SUPPORT
          - MEETING
          - MILESTONE
        priority:
          type: string
          enum:
          - LOW
          - NORMAL
          - HIGH
          - URGENT
        dueDate:
          type: string
          format: date-time
        position:
          type: integer
          format: int32
        completedAt:
          type: string
          format: date-time
        estimatedSeconds:
          type: integer
          format: int64
        createdBy:
          type: string
          format: uuid
        assignedTo:
          type: string
          format: uuid
        assigneeIds:
          type: array
          items:
            type: string
            format: uuid
        parentIds:
          type: array
          items:
            type: string
            format: uuid
        checklistTotal:
          type: integer
          format: int32
        checklistCompleted:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
...
    CreateInternalRequestRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 0
        description:
          type: string
          maxLength: 5000
          minLength: 0
        glpiTicketId:
          type: string
          maxLength: 64
          minLength: 0
        priority:
          type: string
        requestingDepartmentId:
          type: string
          format: uuid
        responsibleDepartmentId:
          type: string
          format: uuid
        desiredDueDate:
          type: string
          format: date-time
        assigneeMembershipIds:
          type: array
          items:
            type: string
            format: uuid
    CreateRequestTasksRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TaskItem'
          minItems: 1
    TaskItem:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 0
        description:
          type: string
          maxLength: 5000
          minLength: 0
        priority:
          type: string
          enum:
          - LOW
          - NORMAL
          - HIGH
          - URGENT
        weight:
          type: integer
          format: int32
        estimatedSeconds:
          type: integer
          format: int64
        dueDate:
          type: string
          format: date-time
        assigneeMembershipIds:
          type: array
          items:
            type: string
            format: uuid
    ConvertRequestToProjectRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 0
        description:
          type: string
          maxLength: 5000
          minLength: 0
    CreateRequestCommentRequest:
      type: object
      properties:
        content:
          type: string
          maxLength: 5000
          minLength: 0
    RequestCommentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        requestId:
          type: string
          format: uuid
        authorMembershipId:
          type: string
          format: uuid
        content:
          type: string
        createdAt:
          type: string
          format: date-time
...
    CreateActivityCommentRequest:
      type: object
      properties:
        content:
          type: string
          maxLength: 4000
          minLength: 0
        mentionMembershipIds:
          type: array
          items:
            type: string
            format: uuid
          maxItems: 20
          minItems: 0
    ActivityCommentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        activityId:
          type: string
          format: uuid
        authorMembershipId:
          type: string
          format: uuid
        authorName:
          type: string
        content:
          type: string
        deleted:
          type: boolean
        canDelete:
          type: boolean
        mentions:
          type: array
          items:
            $ref: '#/components/schemas/ActivityMentionResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ActivityMentionResponse:
      type: object
      properties:
        membershipId:
          type: string
          format: uuid
        displayName:
          type: string
        avatarUrl:
          type: string
...
    ChangeRequestStatusRequest:
      type: object
      properties:
        status:
          type: string
    LinkRequestProjectRequest:
      type: object
      properties:
        projectId:
          type: string
          format: uuid
    LinkRequestActivityRequest:
      type: object
      properties:
        activityId:
          type: string
          format: uuid
    AssignRequestRequest:
      type: object
      properties:
        assigneeMembershipId:
          type: string
          format: uuid
...
    PaginatedResponseInternalRequestResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/InternalRequestResponse'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
...